fix/adjusted-component-names-and-placements-and-fixed-tooltip
