+++ src/components/pages/goals/GoalProgressView.vue
